모든 스테이지 통합
