+++ src/Main.java
@@ -1,9 +1,11 @@
 import common.GameFrame;
+import opening.Opening;
 
 import javax.swing.*;
 
 public class Main extends JFrame{
     public static void main(String[] args) {
-        new GameFrame();
+        new Opening();
+//        new GameFrame();
     }
 }
